added authorization views file
