+++ cs_app/views/__init__.py
@@ -1,1 +1,2 @@
 from .admin_views import *
+from . authorization_views import *
