+++ cs_app/views/admin_views.py
@@ -24,77 +24,6 @@ def main_view(request):
     return HttpResponse(template.render())
 
 
-def create_account_view(request):
-    if request.method == "POST":
-        # Get fields
-        first_name = request.POST.get("firstName")
-        last_name = request.POST.get("lastName")
-        email = request.POST.get("email")
-        password = request.POST.get("password")
-        confirm_password = request.POST.get("confirmPassword")
-
-        # Check if passwords match
-        if password != confirm_password:
-            error_message = "Passwords do not match"
-            return render(
-                request,
-                "create_account.html",
-                {"error_message": error_message, "request": request},
-            )
-
-        # Check if username already exists
-        elif User.objects.filter(username=email).exists():
-            error_message = "Email is already being used."
-            return render(
-                request,
-                "create_account.html",
-                {"error_message": error_message, "request": request},
-            )
-
-        else:
-            # Create new user and save into database
-            user = User.objects.create_user(username=email, password=password)
-            user.first_name = first_name
-            user.last_name = last_name
-            user.email = email
-            user.save()
-            return redirect("/login/")
-
-    return render(
-        request, "create_account.html", {"error_message": None, "request": request}
-    )
-
-
-def login_view(request):
-    if request.method == "POST":
-        form = AuthenticationForm(request, request.POST)
-        if form.is_valid():
-            # Get fields
-            username = form.cleaned_data.get("username")
-            password = form.cleaned_data.get("password")
-            user = authenticate(username=username, password=password)
-            if user is not None:
-                login(request, user)
-                # Redirect to a success page after login
-                return redirect("/home/")
-            else:
-                # Invalid login credentials
-                error_message = "Invalid username or password."
-        else:
-            # Form is not valid
-            error_message = "Invalid form data. Please check the input fields."
-    else:
-        form = AuthenticationForm()
-        error_message = None
-    return render(request, "login.html", {"form": form, "error_message": error_message})
-
-
-@login_required
-def logout_view(request):
-    logout(request)
-    return redirect("login")
-
-
 @login_required
 def home_view(request):
     user = request.user
